% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Almer.R
\name{Almer_boot}
\alias{Almer_boot}
\title{Parametric bootstrap on \code{\link{Almer}} model fit}
\usage{
Almer_boot(mod, nsim = 1000)
}
\arguments{
\item{mod}{fitted object from \code{\link{Almer}}}

\item{nsim}{numder of simulations.}
}
\value{
\code{Almer_boot} a list with entries fixef, vcov, fixef_distribution and vcov_distribution, where the two first entries includes
the means, standard deviations, and quantiles of the fixed effects means and (co)variances, respectively, and the two latter includes all
the bootstrap values.
}
\description{
\code{Almer_boot} Parametric bootstrap on \code{\link{Almer}} model fit
}
\details{
\code{Almer_boot} ...
}
\examples{

}
\author{
Geir H. Bolstad
}
