% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Almer_SE.R
\name{Almer_SE}
\alias{Almer_SE}
\title{!!! NEED TO FIX !!! Linear mixed model with correlated random effects structure}
\usage{
Almer_SE(formula, SE = NULL, maxiter = 100, ...)
}
\arguments{
\item{formula}{as in \code{\link{lmer}}.}

\item{SE}{a vector of standard errors associated with the response variable.}

\item{...}{as in \code{\link{lmer}}.}

\item{A}{as in \code{\link{Almer}}.}

\item{data}{as in \code{\link{lmer}}.}

\item{REML}{as in \code{\link{lmer}}.}

\item{control}{as in \code{\link{lmer}}.}

\item{start}{as in \code{\link{lmer}}.}

\item{verbose}{as in \code{\link{lmer}}.}

\item{subset}{as in \code{\link{lmer}}.}

\item{weights}{as in \code{\link{lmer}}.}

\item{na.action}{as in \code{\link{lmer}}.}

\item{offset}{as in \code{\link{lmer}}.}

\item{contrasts}{as in \code{\link{lmer}}.}

\item{devFunOnly}{as in \code{\link{lmer}}.}
}
\value{
\code{Almer_SE} an object of class \code{\link{merMod}}.
}
\description{
\code{Almer_SE} !!! NEED TO FIX !!! Linear mixed model with correlated random effects structure
}
\examples{

}
\author{
Geir H. Bolstad
}
