% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Almer.R
\name{Almer_sim}
\alias{Almer_sim}
\title{Simulate responses from \code{\link{Almer}} fit}
\usage{
Almer_sim(mod, nsim = 1000)
}
\arguments{
\item{mod}{A fitted object from \code{\link{Almer}}}

\item{nsim}{The number of simulations.}
}
\value{
\code{Almer_sim} a matrix of simulated responses, columns correspond
  to each simulations.
}
\description{
\code{Almer_sim} Simulate responses from an \code{\link{Almer}} model fit.
}
\details{
This function is only included as the \code{\link{simulate.merMod}}
  function did not seem to work properly when the number of random effect
  levels equal the number of observations.
}
\examples{
# See the vignette 'Phylogenetic mixed model'.
}
\author{
Geir H. Bolstad
}
