% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_gls.R
\name{boot_rate_gls}
\alias{boot_rate_gls}
\title{Boostrap of the rate gls model fit}
\usage{
boot_rate_gls(object, n = 10)
}
\arguments{
\item{n}{number of bootsrap samples}

\item{mod}{output from \code{\link{rate_gls}}.}
}
\value{
\code{boot_rate_gls}
}
\description{
\code{boot_rate_gls} Boostrap of the rate gls model fit
}
\details{
\code{boot_rate_gls} Provides a parametric boostrap of the generalized least squares rate model fit 
using \code{\link{simulate.rate_gls}}
}
\examples{

}
\author{
Geir H. Bolstad
}
