\name{evolvability-package}
\alias{evolvability-package}
\docType{package}
\title{
Calculate parameters of evolvability
}
\description{
This package calculates evolvability parameters from estimated G-matrices.
}
\author{Geir H. Bolstad \email{geir.h.bolstad@ntnu.no}}
\references{
Hansen, T. F. & Houle, D. (2008) Measuring and comparing evolvability and constraint in multivariate characters. J. Evol. Biol. 21:1201-1219.
}
