% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanStdG.R
\name{meanStdG}
\alias{meanStdG}
\title{Mean standardize a variance matrix}
\usage{
meanStdG(G, means)
}
\arguments{
\item{G}{A variance matrix.}

\item{means}{A vector of trait means.}
}
\value{
A mean standardized variance matrix.
}
\description{
\code{meanStdG} mean standardizes a variance matrix (e.g. a G-matrix).
}
\examples{
G <- matrix(c(1, 1, 0, 1, 4, 1, 0, 1, 2), ncol = 3)
means <- c(1, 1.4, 2.1)
meanStdG(G, means)
}
\author{
Geir H. Bolstad
}
\keyword{algebra}
\keyword{array}
