% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylH.R
\name{phylH_boot}
\alias{phylH_boot}
\title{Phylogenetic heritability}
\usage{
phylH_boot(mod, numerator, residual = "Residual", nsim = 999)
}
\arguments{
\item{mod}{a mer mode object}

\item{numerator}{name of phylogenetic effect level}

\item{residual}{name of the residual effect level}

\item{nsim}{number of bootstraps}
}
\value{
\code{phylH_boot}
}
\description{
\code{phylH_boot} Phylogenetic heritability from a merMod object
}
\details{
\code{phylH_boot} calculates the phylogenetic heritability with uncertainty using parameteric bootstrapping
}
\examples{

}
\author{
Geir H. Bolstad
}
