% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_gls.R
\name{plot.rate_gls}
\alias{plot.rate_gls}
\title{Plot of gls_rate object}
\usage{
\method{plot}{rate_gls}(object, scale = "SD", main = "GLS regression",
  xlab = "x (mean centred)", ylab = "Response", col = "grey", ...)
}
\arguments{
\item{object}{a gls_rate object}

\item{scale}{either the variance scale ("VAR") or the standard deviation scale ("SD").}

\item{main}{as in \code{\link{plot}}.}

\item{xlab}{as in \code{\link{plot}}.}

\item{ylab}{as in \code{\link{plot}}.}

\item{col}{as in \code{\link{plot}}.}

\item{...}{additional arguments passed to \code{\link{plot}}.}
}
\value{
\code{plot} returns a plot of the gls rate regression
}
\description{
\code{plot} Plots the gls rate regression
}
\details{
\code{plot} ...

Plots the gls rate regression from a gls_rate object obtained from the gls_rate function. The regression line 
gives the expected variance or standard deviation (depending on scale).The regression is linear on the variance scale.
}
\author{
Geir H. Bolstad
}
