% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rate.R
\name{plot.simulate_rate}
\alias{plot.simulate_rate}
\title{Plot of simulate_rate object}
\usage{
\method{plot}{simulate_rate}(
  x,
  response = "rate_y",
  xlab = "Simulation timesteps",
  ylab = "Evolutionary rate of y",
  ...
)
}
\arguments{
\item{x}{An object of class \code{'simulate_rate'}.}

\item{response}{The variable for the y-axis of the plot, can be 'rate_y',
'y', or 'x'.}

\item{xlab}{A label for the x axis.}

\item{ylab}{A label for the y axis.}

\item{...}{Additional arguments passed to \code{\link{plot}}.}
}
\value{
\code{plot} A plot of the evolution of the traits x or y, or the
 evolution of the evolutionary rate of y (i.e. \eqn{\sqrt{a + bx}}) in the
 simulation.
}
\description{
\code{plot} method for class \code{'simulate_rate'}.
}
\details{
No plot is returned if model = 'recent_evol'.
}
\examples{
# See the vignette 'Analyzing rates of evolution'.
}
\author{
Geir H. Bolstad
}
