% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_gls.R
\name{rate_gls}
\alias{rate_gls}
\title{Generalized least squares rate model}
\usage{
rate_gls(x, y, tree, model = "predictor_BM", Beta = as.matrix(c(mean(y^2),
  0)), maxiter = 100, silent = FALSE)
}
\description{
\code{ratesim} Generalized least squares rate model
}
\details{
\code{ratesim} Fits a generalized least squares model to estimate parameters of the evolutoinary model of two traits x and y, 
where the rate of y depends on the value of x. x evolves eiter accoriding to a Brownian motion model of evolution or a geometric 
Brownian motion  model of evolution.
}
