% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_gls.R
\name{rate_gls_boot}
\alias{rate_gls_boot}
\title{Boostrap of the rate gls model fit}
\usage{
rate_gls_boot(object, n = 10, useLFO = TRUE, silent = FALSE)
}
\arguments{
\item{n}{number of bootsrap samples}

\item{useLFO}{when calulating the mean vector of the traits in the "recent_evol" analysis, should the focal species
be left out when calculating the corresponding species' mean. The correct way is to use TRUE, but in practice it has little effect and FALSE 
will speed up the model fit (particularely useful when bootstrapping)}

\item{silent}{}

\item{mod}{output from \code{\link{rate_gls}}.}
}
\value{
A list where the first slot is a table with the original estimates and SE from the GLS fit in the two first columns
followed by the bootstrap estiamte of the SE and the 2.5\%, 50\% and 97.5\% quantiles of the boostrap distribution. The second slot is the complete distribution.
}
\description{
\code{rate_gls_boot} Boostrap of the rate gls model fit
}
\details{
\code{rate_gls_boot} Provides a parametric boostrap of the generalized least squares rate model fit 
using \code{\link{simulate.rate_gls}}
}
\examples{

}
\author{
Geir H. Bolstad
}
