% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_glsboot.R
\name{rate_glsboot}
\alias{rate_glsboot}
\title{Boostrap of the rate gls model fit}
\usage{
rate_glsboot(mod, n = 10)
}
\arguments{
\item{mod}{output from \code{\link{rate_gls}}.}

\item{n}{number of bootsrap samples}
}
\value{
\code{rate_glsboot}
}
\description{
\code{rate_glsboot} Boostrap of the rate gls model fit
}
\details{
\code{rate_glsboot} Provides a parametric boostrap of the generalized least squares rate model fit, using \code{\link{rate_sim}}
}
\examples{

}
\author{
Geir H. Bolstad
}
