% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_sim.R
\name{rate_sim}
\alias{rate_sim}
\title{Simulating evolutionary rate model}
\usage{
rate_sim(tree, startv_x = NULL, sigma_x = NULL, a, b, x = NULL,
  model = "predictor_BM")
}
\arguments{
\item{tree}{a \code{\link{phylo}} object. Must be ultrametric and scaled to unit length.}

\item{sigma_x}{evolutionary rate of x}

\item{a}{parameter of the evolutionary rate of y (see details)}

\item{b}{parameter of the evolutionary rate of y (see details)}

\item{x}{fixed values of x (only for the "residual_rate" model), 
must equal number of tips in the pylogeny, must be the same order as the tip labels}

\item{model}{either Brownaian motion model of x "predictor_BM", 
geometric Brownian motion model of x "predictor_geometricBM", or "residual_rate".}

\item{starv_x}{star value for x}
}
\value{
\code{ratesim} returns a dataset of x and y values at the tips.
}
\description{
\code{rate_sim} Simulating evolutionary rate model
}
\details{
\code{rate_sim} Simulates different evolutionary rates models where a trait x evolves according to a Brownian motion or geometric Bronian motion 
process (constant evolutionary rate) while the rate of a second trait, y, is a linear function of x.
}
\examples{

}
\author{
Geir H. Bolstad
}
