% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_and_format.R
\name{round_and_format}
\alias{round_and_format}
\title{Rounds and formats in the same function}
\usage{
round_and_format(x, digits = 2, scientific = FALSE, trim = TRUE)
}
\arguments{
\item{x}{data in numeric format}

\item{digits}{Number of decimals}

\item{scientfic}{Logical specifying whether encoding in scientific notation}
}
\value{
rounded and formatted values
}
\description{
\code{round_and_format} rounds and formats data
}
\author{
Geir H. Bolstad
}
