% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_gls.R
\name{simulate.rate_gls}
\alias{simulate.rate_gls}
\title{Simulate responses from \code{\link{rate_gls}} fit}
\usage{
\method{simulate}{rate_gls}(object, nsim = 10)
}
\arguments{
\item{object}{fitted object from \code{\link{rate_gls}}}

\item{nsim}{numder of simulations.}
}
\value{
\code{simulate.rate_gls} returns a list where each slot is an interation.
}
\description{
\code{simulate.rate_gls} Simulate responses from \code{\link{rate_gls}} fit
}
\details{
\code{simulate.rate_gls} ...
}
\examples{

}
\author{
Geir H. Bolstad
}
