% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_Almer.R
\name{simulate_Almer}
\alias{simulate_Almer}
\title{Simulate responses from \code{\link{Almer}} fit}
\usage{
simulate_Almer(mod, nsim = 10)
}
\arguments{
\item{mod}{fitted object from \code{\link{Almer}}}

\item{nsim}{numder of simulations.}
}
\value{
\code{simulate_Almer} a matrix of simulated responses, colunms correspond to each simulations.
}
\description{
\code{simulate_Almer} Simulate responses from \code{\link{Almer}} fit
}
\details{
\code{simulate_Almer} ...

This function is only included as the simulate.merMod funciton did not seem to work properly when the 
number of random effect levels equal the number of observations.
}
\examples{

}
\author{
Geir H. Bolstad
}
