% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rate.R
\name{simulate_rate}
\alias{simulate_rate}
\title{Simulating evolutionary rate model}
\usage{
simulate_rate(
  tree,
  startv_x = NULL,
  sigma_x = NULL,
  a,
  b,
  sigma_y = NULL,
  x = NULL,
  model = "predictor_BM"
)
}
\arguments{
\item{tree}{A \code{\link{phylo}} object. Must be ultrametric and scaled to unit length.}

\item{startv_x}{The starting value for x (usually 0 for "predictor_BM" and "recent_evol", and 1 for "predictor_gBM").}

\item{sigma_x}{The evolutionary rate of x.}

\item{a}{A parameter of the evolutionary rate of y.}

\item{b}{A parameter of the evolutionary rate of y.}

\item{sigma_y}{The evolutionary rate for the macroevolution of y (Brownian motion process) used in the "recent_evolution" model.}

\item{x}{optional fixed values of x (only for the "recent_evol" model), 
must equal number of tips in the pylogeny, must correspond to the order of the tip labels.}

\item{model}{either Brownaian motion model of x "predictor_BM", 
geometric Brownian motion model of x "predictor_gBM", or "recent_evol".}
}
\value{
\code{simulate_rate} returns a data frame of x and y values for the tips.
}
\description{
\code{simulate_rate} Simulates three different evolutionary rates models. In the two first, 
"predictor_BM" and "predictor_gBM", the evolution of y follows a Brownian motion 
with variance linear in x, while the evolution of x either follows a Brownian 
motion or a geometric Brownian motion, respectively. In the third model, "recent_evol", the 
residuals of the macroevolutionary predictions of y have variance linear in x.
}
\details{
See the vignette "Analysing rates of evolution" for examples of use. A large part of parameter space will cause negative 
rates for y (i.e. negative a+bx). In these cases the rates are set to 0. A warning message is given if the number of such instances 
is larger than 0.01%.
}
\examples{
# See the vignette "Analysing rates of evolution".
}
\author{
Geir H. Bolstad
}
