% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Almer.R
\name{Almer_SE}
\alias{Almer_SE}
\title{Linear mixed model for response variables with uncertainty}
\usage{
Almer_SE(
  formula,
  SE = NULL,
  maxiter = 100,
  control = lme4::lmerControl(check.nobs.vs.nlev = "ignore", check.nobs.vs.rankZ =
    "ignore", check.nobs.vs.nRE = "ignore"),
  ...
)
}
\arguments{
\item{formula}{as in \code{\link{lmer}}.}

\item{SE}{A vector of standard errors associated with the response variable. NB! Must have column name "SE" in the data.}

\item{maxiter}{The maximum number of iterations.}

\item{control}{as in \code{\link{lmer}}.}

\item{...}{Further optional arguments, see \code{\link{Almer}}.}
}
\value{
\code{Almer_SE} returns an object of class \code{\link{merMod}}.
}
\description{
\code{Almer_SE} Linear mixed model for response variables with uncertainty
}
\examples{
# See the vignette 'Phylogenetic mixed model'.
}
\author{
Geir H. Bolstad
}
