% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Almer_SE.R
\name{Almer_SE}
\alias{Almer_SE}
\title{Linear mixed model for responsvariables with uncertainty}
\usage{
Almer_SE(formula, SE = NULL, maxiter = 100, ...)
}
\arguments{
\item{formula}{as in \code{\link{lmer}}.}

\item{SE}{a vector of standard errors associated with the response variable.}

\item{maxiter}{maximum number of iterations.}

\item{...}{optional arguments, see \code{\link{Almer}}.}
}
\value{
\code{Almer_SE} an object of class \code{\link{merMod}}.
}
\description{
\code{Almer_SE} Linear mixed model for responsvariables with uncertainty
}
\examples{

}
\author{
Geir H. Bolstad
}
