\name{evolvability-package}
\alias{evolvability-package}
\alias{evolvability}
\docType{package}
\title{
Calculation of Evolvability Parameters
}
\description{
This package calculates evolvability parameters from estimated G-matrices as defined in Hansen and Houle (2008) <doi:10.1111/j.1420-9101.2008.01573.x>. It can use both point estimates and posterior/bootstrap distributions of the G-matrices. This package was released with Bolstad et al. (2014) <doi:10.1098/rstb.2013.0255>, which contains some examples of use.
}
\details{
\tabular{ll}{
Package: \tab evolvability\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2015-04-13\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Geir H. Bolstad \email{geir.bolstad@nina.no}}
\references{
Bolstad G. H., Hansen T. F., Pelabon C. Falahati-Anabaran M., Perez-Barrales R. & Armbruster W. S. 2014. Genetic constraints predict evolutionary divergence in Dalechampia blossoms. Phil. Trans. R. Soc. B. 369:20130255. doi:10.1098/rstb.2013.0255

Hansen T. F. & Houle D. (2008) Measuring and comparing evolvability and constraint in multivariate characters. J. Evol. Biol. 21:1201-1219. doi:10.1111/j.1420-9101.2008.01573.x
}
\keyword{package}
