\name{evolvability-package}
\alias{evolvability-package}
\alias{evolvability}
\docType{package}
\title{
Calculate parameters of evolvability
}
\description{
This package calculates evolvability parameters from estimated G-matrices. It can use both point estimates and posterior/bootstrap distributions of variance matrices.
}
\details{
\tabular{ll}{
Package: \tab evolvability\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2015-04-13\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Geir H. Bolstad \email{geir.bolstad@nina.no}}
\references{
Hansen, T. F. & Houle, D. (2008) Measuring and comparing evolvability and constraint in multivariate characters. J. Evol. Biol. 21:1201-1219.
}
\keyword{ package }
