% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylH.R
\name{phylH}
\alias{phylH}
\title{Phylogenetic heritability}
\usage{
phylH(mod, numerator, residual = "Residual", nsim = 10)
}
\arguments{
\item{mod}{An object of class \code{'merMod'}}

\item{numerator}{The name of phylogenetic effect level}

\item{residual}{name of the residual effect level}

\item{nsim}{number of bootstraps}
}
\value{
\code{phylH} returns a list with the REML estimate, the 95\%
  confidence interval from the parametric bootstrap, and the bootstrap
  samples.
}
\description{
\code{phylH} calculates the phylogenetic heritability from an \code{Almer}
model fit and provides associated uncertainty using parametric
bootstrapping.
}
\examples{
# See the vignette 'Phylogenetic mixed model'.
}
\author{
Geir H. Bolstad
}
