% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rate.R
\name{simulate_rate}
\alias{simulate_rate}
\title{Simulating evolutionary rate model}
\usage{
simulate_rate(
  tree,
  startv_x = NULL,
  sigma_x = NULL,
  a,
  b,
  sigma_y = NULL,
  x = NULL,
  model = "predictor_BM"
)
}
\arguments{
\item{tree}{a \code{\link{phylo}} object. Must be ultrametric and scaled to unit length.}

\item{sigma_x}{evolutionary rate of x.}

\item{a}{parameter of the evolutionary rate of y (see details).}

\item{b}{parameter of the evolutionary rate of y (see details).}

\item{sigma_y}{evolutionary rate for the macroevolution of y (Brownian motion process) used in the "recent_evolution" model.}

\item{x}{optional fixed values of x (only for the "recent_evol" model), 
must equal number of tips in the pylogeny, must correspond to the order of the tip labels}

\item{model}{either Brownaian motion model of x "predictor_BM", 
geometric Brownian motion model of x "predictor_gBM", or "recent_evol".}

\item{starv_x}{starting value for x.}
}
\value{
\code{simulate_rate} returns a dataset of x and y values for the tips.
}
\description{
\code{simulate_rate} Simulating evolutionary rate model
}
\details{
\code{simulate_rate} Simulates different evolutionary rates models where a trait x evolves according to a Brownian motion or geometric Bronian motion 
process (constant evolutionary rate) while the rate of a second trait, y, is a linear function of x.
}
\examples{

}
\author{
Geir H. Bolstad
}
