% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolvabilityBeta.R
\name{summary.evolvabilityBeta}
\alias{summary.evolvabilityBeta}
\title{Summarizing evolvability parameters over a set of selection gradients}
\usage{
\method{summary}{evolvabilityBeta}(object, ...)
}
\arguments{
\item{object}{An object of class \code{'evolvabilityBeta'}.}

\item{...}{Additional arguments.}
}
\value{
A list with the following components:
\tabular{llllll}{
\code{Averages} \tab\tab\tab\tab The averages of the evolvability parameters 
over all selection gradients. \cr
\code{Minimum} \tab\tab\tab\tab The minimum of the evolvability parameters 
over all selection gradients. \cr
\code{Maximum} \tab\tab\tab\tab The maximum of the evolvability parameters 
over all selection gradients.
}
}
\description{
\code{summary} method for class \code{'evolvabilityBeta'}.
}
\seealso{
\code{\link{evolvabilityBeta}}
}
\author{
Geir H. Bolstad
}
\keyword{algebra}
\keyword{array}
